config/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-rest-passwordreset==1.3.0
django-simple-history==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
